+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: jstomps <jstomps@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/10/08 16:58:55 by jstomps       #+#    #+#                 */
-/*   Updated: 2025/10/09 16:27:15 by jstomps       ########   odam.nl         */
+/*   Updated: 2025/10/14 18:29:15 by jstomps       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while ((s1[i] == s2[i] && i < n) && (s1 && s2))
